docs: Readme
Copied from the coingeco website

On September 4th, MATIC will transition to POL. In the initial phase, POL will replace MATIC as the native gas and staking token for Polygon PoS, with subsequent phases expanding POL’s role in the AggLayer. For more information, please refer to this post. For the new migrated token, visit POL on CoinGecko.